staging: fsl-dpaa2: rtc: add rtc driver
This patch is to add DPAA2 RTC driver which is PTP 1588 clock
driver for DPAA2. The driver registers itself with class driver
for using the PTP hardware clock infrastructure. Together with
the SO_TIMESTAMPING socket options, this presents a standardized
method for developing PTP user space programs, synchronizing
Linux with external clocks, and using the ancillary features of
PTP hardware clocks.

Signed-off-by: Yangbo Lu <yangbo.lu@nxp.com>