spi: spi-fsl-dspi: replace regmap R/W with internal implementation
There are some number of problems in the course of actual use of
the regmap which is designed very large and complex.

For DSPI driver we just only want to R/W the controller register
with BE or LE.

So I can just packaging a sample function which can fully capable
of this work.

Signed-off-by: Yuan Yao <yao.yuan@nxp.com>