net: fec: enable RX flushing for Best Effort queue
Before packets are dispatched to individual RX queues, the packets are
stored in a common queue. If one of the queues is blocked because the
driver cannot process the packets from that queue then the other queues
will be blocked as well.

For AVB traffic RX flushing is required to prevent class A (queue 1)
and class B (queue 2) from being blocked by the Best Effort class
(queue 0).

Also for the TX QoS scheme the credit based shaper is enabled by
default as its value is 0x0.

Signed-off-by: Marouen Ghodhbane <marouen.ghodhbane@nxp.com>
Signed-off-by: Hou Zhiqiang <Zhiqiang.Hou@nxp.com>