staging: fsl-dpaa2/eth: Add support for Tx congestion
Configure the hardware to write notifications to memory
when Tx queues enter or exit congestion.

On Tx routine, we check if we are in a congested state
and if yes, instruct the stack to stop the current Tx queue.
Queues are restarted during Tx conf processing if the
hardware indicates we're no longer congested.

Add an ethtool stat entry for current congestion status.

Signed-off-by: Ioana Radulescu <ruxandra.radulescu@nxp.com>