spi: spi-fsl-dspi: replace regmap R/W with internal
There are some number of problems in the course of actual use of
the regmap which is designed very large and complex.

For DSPI driver we just only want to R/W the controller register
with BE or LE.

So I can just packaging a sample function which can fully capable
of this work.

Signed-off-by: Zhang Ying-22455 <ying.zhang22455@nxp.com>