staging: fsl-dpaa2/eth: Release buffers on XDP_DROP
Instead of freeing the Rx FD buffer, release it
back into the pool. We wait for the maximum number
of buffers (7) supported by a single release command
to accumulate before issuing the command.

Also, don't unmap the Rx buffers at the beginning of
the Rx routine anymore, since remapping them before
release adds an unnecessary overhead. Instead, just
do a DMA sync at first and only unmap if the frame
is meant for the stack.

Signed-off-by: Ioana Radulescu <ruxandra.radulescu@nxp.com>