staging: fsl-dpaa2/eth: Don't enable FAS on Tx
For Tx confirmed frames, we look at the Frame Annotation Status
field of error frames for indication on the error cause and then
print a debug message with that information.

While useful in initial development stages, it doesn't bring
enough added value to justify reserving 64B of headroom for all
Tx frames (FAS is only 8B long, but we must reserve from the
hardware annotation area chunks of 64B).

If we remove the need for FAS field from egress frames, we can
renounce hardware annotation completely, except for frames which
have timestamp enabled.

Signed-off-by: Ioana Radulescu <ruxandra.radulescu@nxp.com>