staging: fsl-dpaa2/eth: Configure HWA on XDP_TX FD
Explicitly reserve a hardware annotation area of 128B in the XDP_TX FD,
since we write to the HWA[FAEAD] field to disable Tx confirmations.

We don't actually use the software annotation area, but since we reuse
the Rx FD and buffer, let the hardware know the SWA is reserved.

Signed-off-by: Ioana Radulescu <ruxandra.radulescu@nxp.com>