enetc: Remove Tx cleanup restriction on complete frames
Cleanup of Tx BDs is possible even if the cleaned-up BDs
don't comprise full frames. This is less restrictive
than forcing cleanup of full frames, and more robust, as
the previous design would end up in an infinite while loop
in case the h/w CISR did not always indicate the start
of a full frame (like in the m0168 simulator model).
So, better to guard the driver against fuzzy h/w specifications,
and use this simplified and more robust Tx cleanup code.

Signed-off-by: Claudiu Manoil <claudiu.manoil@nxp.com>