crypto: caam/jr - fix S/G table passing SMMU 4K page boundary
According to CAAM RM, crypto engine reads 4 S/G entries at a time,
even if the S/G table has fewer entries; it's the responsibility of the
user / programmer to make sure this HW behaviour has no side effect.

The driver does not take care of this currently, leading to SMMU faults
when the S/G table is passing the 4k page boundary and only the first
page is DMA mapped.

Fix this by always rounding up the number of S/G table entries to be a
multiple of 4.

Note: ablkcipher with IV generation code is not fixed; this is dead code
and shall be removed in near future.

Signed-off-by: Horia Geantă <horia.geanta@nxp.com>