net: can: flexcan: move rx offload_add from probe to start
rx offload depends on number of message buffers, which in turn depends on
messgae buffer size. with the upcoming LX2160A SOC the message buffer size
can be configured to 72 bytes if it were to be used in CAN FD mode.

The current mode in which the flexcan is being operated is known at the
time of flexcan_open and not at the time of flexcan_probe.

Therefore, move the rx offload_add from flexcan_probe to flexcan_open.
correspondingly, move rx offload_delete form flexcan_remove to
flexcan_close.

Signed-off-by: Pankaj Bansal <pankaj.bansal@nxp.com>