felix: Add packet injection extraction mechanism
This mechanism allows a faster alternative to send/receive frames
than current Ocelot implementation based on register access.
The new mechanism is based on the Felix capability to allow
frames tagged with a 32-byte header to bypass the switch frame
analyzer on xmit. For receive, we use another Felix capability
that enables to forward all frames to the CPU port and have
the frames tagged with 32-byte header from which we can extract
the source port.

Signed-off-by: Catalin Horghidan <catalin.horghidan@nxp.com>