staging: fsl-dpaa2/mac: fix the if_mode type
The of_get_phy_mode function returns a signed integer and should be
treated as such. Since the underlying type for the enum cannot be
determined (that's up to the compiler), it can be signed or unsigned,
there may be cases when the comparison to 0 is always true.

Fix this by making the if_mode type integer.

should be signed since of_get_phy_mode could return negative
also

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>