sdk: dts: Update dma-coherent settting on ls104x qds sd and rdb-usdpaa
This is supplemental patch of ("Move dma-coherent form soc to its
child nodes"). Add support for qds and rdb-usdpaa dts files.

Since SMMU is not supported for SDK version, USB function will down
if still apply property 'dma-coherent' in scope of soc (USB driver
is not ready to support it alone) in SDK device trees,decide to remove it.
And add dma-coherent on other non-USB child nodes under soc.

Signed-off-by: Ran Wang <ran.wang_1@nxp.com>