felix: add injection timestamping support
Injection timestamping should have been implemented in ISR
of injection or ptp timestamping ready, but we were facing
problem to trigger the switch interrupt.

This patch is to use polling in work queue after injection
to handle timestamping instead as a workaround. This is
temporary until the switch interrupt is implemented.

Signed-off-by: Yangbo Lu <yangbo.lu@nxp.com>