ocelot: Warn if cpu side interfaces are not up
For correct usage of external switch ports the cpu
side ethernet interfaces - the switch's NPI port and
the pair etsec nedevice connected to it - must be brought
up first before bringing up any external switch port.
However the switchdev model doesn't enforce this restriction
and implementing it in the driver would require considerable
rework (not practical at this point).

So best we can do is to warn the user that the CPU side
interfaces need to be brought up manually. If not, L2
fwd-ing among external switch ports may work, but the
host has no way to know that it cannot use the external
switch ports for transmission and so it will attempt
to do so and hog the cpu by calling port xmit functions
that cannot send packets.

Signed-off-by: Claudiu Manoil <claudiu.manoil@nxp.com>