staging: fsl-dpaa2/mac: do not call dpmac_set_link_state() when nothing changed
In case nothing changed in the link configuration do not call
dpmac_set_link_state().
This is needed in case of the following sequence of commands.

$ ip link set dev eth1 up; ip link set dev eth2 down

Phylib brings the link down when the aneg is started on the phy which
translates in a link down from phy in MC and confuses the MC linkman.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>